+++ system/src/Grav/Framework/Media/Interfaces/MediaCollectionInterface.php
@@ -3,7 +3,7 @@
 /**
  * @package    Grav\Framework\Media
  *
- * @copyright  Copyright (C) 2015 - 2019 Trilby Media, LLC. All rights reserved.
+ * @copyright  Copyright (C) 2015 - 2020 Trilby Media, LLC. All rights reserved.
  * @license    MIT License; see LICENSE file for details.
  */